requirements.txt
Pillow>=8.0.0
requests>=2.25.0
python-vlc>=3.0.0
psutil>=5.8.0
qrcode>=7.0 